backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-multipart==0.0.6
pydantic==2.5.0
openai-whisper==20231117
torch==2.1.0
sentence-transformers==2.2.2
numpy==1.24.3
python-jose[cryptography]==3.3.0
